P-Store: A read-optimized database system
P-Store is a simple read-optimized database system. The command line tool
"pstore" implements two commands: "import" and "cat". The former can be used to
convert a CSV file into a P-Store database and the latter can be used to
examine contents of the database.

The design of this system is loosely based on C-Store which is another Open
Source read-optimized database. The main idea is to store tabular data in
column-order to improve read performance for special purpose workloads.

P-Store is in very early stages so don't expect a whole lot from it.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>