Use sliding-window mmap()'d memory to read CSV files
Buffered I/O is the bottleneck of "pstore import". Unfortunately we can't just
mmap() the whole file because that limits us to 2 GB files on 32-bit systems.
To support larger files, implement a "sliding-window mmap()" where we mmap() a
fixed region of 128 MB and move the window forward until we've gone through the
whole file.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>