mmap-window: Fix position in newly mapped region
There's a bug in mmap_window__slide() for how we calculate the position in the
mmap()'d region. The memory mapped before self->pos (due to mmap() page
alignment requirement) is not considered part of the window but self->mmap
points to the physical start of the mmap()'d region.

Therefore make sure we take self->mmap_pos into account when calculating the
current "absolute" position in the sliding mmap() region.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>