extent: Make PSTORE_LAST_EXTENT constant explicit
We need a magic number for on-disk ->next_extent that indicates the last
extent. Currently that value is implicitly zero which is fine as offset zero is
always reserved for the header and thus can't be part of any extent.

To make things more explicit for future code, introduce a PSTORE_LAST_EXTENT
constant and use it.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>