extents: Change on-disk layout to support compression
This patch changes the on-disk extent header layout to support compression.

After this change, each extent has two different sizes: lsize and psize. The
'lsize' field represents the logical size without compression (in-memory size)
and the 'psize' field represents physical size after compression (on-disk
size). A new 'comp' field is also added that records the compression type used
for an extent.

For uncompressed extents, both 'lsize' and 'psize' are equal and 'comp' is set
to zero.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>