Fix maximum window size check in mmap_window__mmap()
We should always include the mmap alignment in mmap length; otherwise we might
never reach the end of the mmap region. The problem is not just theoretical, we
already have some questionable code in pstore_extent__decompress() to work
around the problem. You can also see the problem with 'pstore import' if you
force maximum extent size and window length to be equal size:

  $ ./pstore import --max-extent-len=1 --compress --window-len=1 ~/testing/pstore/aapl-small.csv tmp
  csv_iterator_next: premature end of file

This patch fixes both issues.

Signed-off-by: Pekka Enberg <penberg@cs.helsinki.fi>