disk-format: Fix struct pstore_file_extent 64-bit issues
The ->padding member in struct pstore_file_extent pads to 4 bytes but on 64-bit
the required alignment is 8 bytes. This means that on 32-bit we have:

  (gdb) print sizeof(struct pstore_file_extent)
  $1 = 28

and on 64-bit we have:

  (gdb) print sizeof(struct pstore_file_extent)
  $1 = 32

which means that ABI is different on 32-bit and 64-bit. This patch fixes the
issue by forcing 64-bit compatible padding. Unfortunately that also means old
32-bit files don't work anymore.

Signed-off-by: Pekka Enberg <penberg@kernel.org>