builtin-import: Add '--append' option
The '--append' option appends the contents of a CSV file into an
existing P-Store database.

New extents are written to the end of the file and linked to the
old extents. In order to reduce file traversal, the disk format
is updated: a 'last_extent' field is added to the column header.
Unfortunately that means old files don't work anymore.

Signed-off-by: Jussi Virtanen <jussi.virtanen@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@kernel.org>