builtin-repack: Fix multi-extent segment offsets
There was a fault in builtin-repack when working on files that
result in multi-extent segments in repacking. The fault was such
that 'next_extent' fields in extent headers were not pointing to
the following extents.

The fault is fixed by updating the 'column->prev_extent' field,
which is later on accessed in 'pstore_column__flush_write'.

It is not clear which commit has caused the fault. However, at
least the following commit is very likely related to the matter:

  0897e9a7b48a1a754ea2ab6867747983314048df

Signed-off-by: Jussi Virtanen <jussi.virtanen@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@kernel.org>