builtin-extend: Preallocate extents
Add a new 'pstore extend' command that preallocates extents
for segments. Preallocated extents are uncompressed.

By default, 'pstore extend' preallocates a 16 MiB extent for
all columns in all tables. The '--column' and '--table' options
limit the operation to specific columns in specific tables. The
extent length can be specified with the '--extent-len' option.

The on-disk layout of a P-Store database does not change, but
for uncompressed extents the 'lsize' now tells how much of the
storage space is in use and the 'psize' how much storage space
there is.

On its own, 'pstore extend' is useless. However, it is groundwork
for enhancing 'pstore import --append' to fill up the last extent
of a segment before appending a new extent.

Signed-off-by: Jussi Virtanen <jussi.virtanen@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@kernel.org>