extents: Add support for Snappy
Snappy is a fast compression/decompression library from Google.

The support for Snappy is optional: if the environment variable
$SNAPPY_HOME is defined, P-Store is compiled with Snappy support
and linked with 'libsnappy.a', otherwise P-Store is compiled
without Snappy support.

Here is a comparison of FastLZ and Snappy with a 1 GiB CSV file
generated by 'torture/gencsv':

    FastLZ                 Snappy

    import
      real  0m15.634s      0m16.675s
      user  0m9.868s       0m10.621s
      sys   0m2.537s       0m2.630s

    export
      real  0m17.847s      0m17.013s
      user  0m6.279s       0m5.963s
      sys   0m2.301s       0m2.442s

    compression ratio
            33.64%         40.35%

Snappy was built with default options.

Signed-off-by: Jussi Virtanen <jussi.virtanen@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@kernel.org>