Add '--format' option to 'pstore import'
Add a '--format' option to 'pstore import' as an alternative to
the '--delimiter' and '--quote' options. It makes setting the input
format to TSV easier than by changing the delimiter.

Besides TSV, the '--format' option supports CSV. Unlike the default
input format, this input format support quoting. In other words, it
is equivalent to setting '--quote' to '"'. (Quoting is disabled by
default because it incurs a performance penalty.)

Signed-off-by: Jussi Virtanen <jussi.k.virtanen@gmail.com>